Adding an abstract TranslationEngine base class for translations.
