+++ src/ORM/Behavior/Translate/EavEngine.php
@@ -18,12 +18,9 @@ namespace Cake\ORM\Behavior\Translate;
 use ArrayObject;
 use Cake\Collection\Collection;
 use Cake\Event\Event;
-use Cake\I18n\I18n;
-use Cake\Core\InstanceConfigTrait;
 use Cake\ORM\Behavior;
 use Cake\ORM\Entity;
 use Cake\ORM\Query;
-use Cake\ORM\Table;
 use Cake\ORM\TableRegistry;
 
 /**
@@ -38,67 +35,11 @@ use Cake\ORM\TableRegistry;
  * If you want to bring all or certain languages for each of the fetched records,
  * you can use the custom `translations` finders that is exposed to the table.
  */
-class EavEngine
+class EavEngine extends TranslationEngine
 {
 
-    use InstanceConfigTrait;
-
     /**
-     * Table instance
-     *
-     * @var \Cake\ORM\Table
-     */
-    protected $_table;
-
-    /**
-     * The locale name that will be used to override fields in the bound table
-     * from the translations table
-     *
-     * @var string
-     */
-    protected $_locale;
-
-    /**
-     * Instance of Table responsible for translating
-     *
-     * @var \Cake\ORM\Table
-     */
-    protected $_translationTable;
-
-    /**
-     * Default config
-     *
-     * These are merged with user-provided configuration when the behavior is used.
-     *
-     * @var array
-     */
-    protected $_defaultConfig = [
-        'implementedFinders' => ['translations' => 'findTranslations'],
-        'implementedMethods' => ['locale' => 'locale'],
-        'fields' => [],
-        'translationTable' => 'I18n',
-        'defaultLocale' => '',
-        'model' => '',
-        'onlyTranslated' => false,
-        'strategy' => 'subquery',
-        'conditions' => ['model' => '']
-    ];
-
-    /**
-     * Constructor
-     *
-     * @param \Cake\ORM\Table $table The table this behavior is attached to.
-     * @param array $config The config for this behavior.
-     */
-    public function __construct(Table $table, array $config = [])
-    {
-        $config += ['defaultLocale' => I18n::defaultLocale()];
-        $this->config($config);
-        $this->_table = $table;
-    }
-
-    /**
-     * Initialize hook
+     * {@inheritDoc}
      *
      * @param array $config The config for this behavior.
      * @return void
@@ -183,9 +124,7 @@ class EavEngine
     }
 
     /**
-     * Callback method that listens to the `beforeFind` event in the bound
-     * table. It modifies the passed query by eager loading the translated fields
-     * and adding a formatter to copy the values into the main table records.
+     * {@inheritDoc}
      *
      * @param \Cake\Event\Event $event The beforeFind event that was fired.
      * @param \Cake\ORM\Query $query Query
@@ -247,8 +186,7 @@ class EavEngine
     }
 
     /**
-     * Modifies the entity before it is saved so that translated fields are persisted
-     * in the database too.
+     * {@inheritDoc}
      *
      * @param \Cake\Event\Event $event The beforeSave event that was fired
      * @param \Cake\ORM\Entity $entity The entity that is going to be saved
@@ -304,7 +242,7 @@ class EavEngine
     }
 
     /**
-     * Unsets the temporary `_i18n` property after the entity has been saved
+     * {@inheritDoc}
      *
      * @param \Cake\Event\Event $event The beforeSave event that was fired
      * @param \Cake\ORM\Entity $entity The entity that is going to be saved
@@ -316,38 +254,7 @@ class EavEngine
     }
 
     /**
-     * Sets all future finds for the bound table to also fetch translated fields for
-     * the passed locale. If no value is passed, it returns the currently configured
-     * locale
-     *
-     * @param string|null $locale The locale to use for fetching translated records
-     * @return string
-     */
-    public function locale($locale = null)
-    {
-        if ($locale === null) {
-            return $this->_locale ?: I18n::locale();
-        }
-        return $this->_locale = (string)$locale;
-    }
-
-    /**
-     * Custom finder method used to retrieve all translations for the found records.
-     * Fetched translations can be filtered by locale by passing the `locales` key
-     * in the options array.
-     *
-     * Translated values will be found for each entity under the property `_translations`,
-     * containing an array indexed by locale name.
-     *
-     * ### Example:
-     *
-     * ```
-     * $article = $articles->find('translations', ['locales' => ['eng', 'deu'])->first();
-     * $englishTranslatedFields = $article->get('_translations')['eng'];
-     * ```
-     *
-     * If the `locales` array is not passed, it will bring all translations found
-     * for each record.
+     * {@inheritDoc}
      *
      * @param \Cake\ORM\Query $query The original query to modify
      * @param array $options Options
